requirements.txt
flask
pymupdf
pillow
python-docx
